moved all engagement-specific content into `/data/engagement.json`
All the jade/scss will compile without this, but if you want to compile
with this content, you should run `./compile_local.sh`

Also, note that the “tags” attribute appears to be incomplete. The full
list of tags and buckets are derived automatically based on
buckets/tags referenced in the list of deliverables. The “tags” array
is essentially used to override default values. For instance, each tag
“slug” (what’s used in the CSS selector) is assumed to be the
lower-/upper-case tag name for tags/buckets, respectively. Intelligent
defaults are used unless otherwise stated in this file.